+++ R/app-server-fit-model.R
@@ -405,6 +405,10 @@ registerFitModelObservers = function(input, output, session, rv, modelFit, reset
       model = m,
       followupPayload = followupClassification
     )
+    followupClassification = attachQuestionRouteToModelFollowupPayload(
+      followupQuestion = followupQuestion,
+      followupPayload = followupClassification
+    )
     attr(m, "wmfm_model_followup_question") = followupClassification$originalText
     attr(m, "wmfm_model_followup_payload") = followupClassification
     promptPreview = lmToExplanationPrompt(m)